moved visualization methods to utils
