--- tent_pitching/visualization/__init__.py
@@ -1,2 +1,0 @@
-from .visualization_1d import (plot_1d_space_time_grid, plot_space_function,
-                               plot_space_time_function, plot_on_reference_tent)
